+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [1.5.3] - 2026-07-23
+
 ## [1.5.2] - 2026-07-23
 
 ## [1.5.1] - 2026-07-23